requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.6
djangorestframework==3.12.2
pytz==2021.1
sqlparse==0.4.1
